backend/pyproject.toml
[project]
name = "modellab-backend"
version = "0.1.0"
description = "ModelLab Backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "pydantic>=2.6.0",
    "pandas>=2.2.0",
    "interpretml>=0.5.1",
    "python-multipart>=0.0.9",
    "pyyaml>=6.0",
    "lightgbm>=4.0.0",
    "optuna>=3.5.0",
    "shap>=0.45.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
